backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# OpenAI and AI
openai>=1.30.0

# Video/Audio processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
yt-dlp==2023.11.16

# Data handling
pydantic==2.5.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: Enhanced video processing
# imageio-ffmpeg==0.4.9
# static-ffmpeg==2.5